Move zero balances filter
