+++ feature-assets/src/main/java/io/novafoundation/nova/feature_assets/presentation/balance/list/BalanceListViewModel.kt
@@ -216,10 +216,6 @@ class BalanceListViewModel(
         router.openSwitchWallet()
     }
 
-    fun filtersClicked() {
-        router.openAssetFilters()
-    }
-
     fun manageClicked() {
         router.openManageTokens()
     }
