+++ feature-assets/src/main/java/io/novafoundation/nova/feature_assets/presentation/tokens/manage/ManageTokensFragment.kt
@@ -15,15 +15,18 @@ import io.novafoundation.nova.common.utils.keyboard.hideSoftKeyboard
 import io.novafoundation.nova.common.utils.setVisible
 import io.novafoundation.nova.common.utils.keyboard.showSoftKeyboard
 import io.novafoundation.nova.common.utils.submitListPreservingViewPoint
+import io.novafoundation.nova.common.view.bindFromMap
 import io.novafoundation.nova.feature_assets.R
 import io.novafoundation.nova.feature_assets.di.AssetsFeatureApi
 import io.novafoundation.nova.feature_assets.di.AssetsFeatureComponent
+import io.novafoundation.nova.feature_assets.domain.assets.filters.NonZeroBalanceFilter
 import kotlinx.android.synthetic.main.fragment_manage_tokens.manageTokensContainer
 import kotlinx.android.synthetic.main.fragment_manage_tokens.manageTokensList
 import kotlinx.android.synthetic.main.fragment_manage_tokens.manageTokensPlaceholder
 import kotlinx.android.synthetic.main.fragment_manage_tokens.manageTokensSearch
 import kotlinx.android.synthetic.main.fragment_manage_tokens.manageTokensToolbar
 import javax.inject.Inject
+import kotlinx.android.synthetic.main.fragment_manage_tokens.manageTokensSwitchZeroBalances
 
 class ManageTokensFragment :
     BaseFragment<ManageTokensViewModel>(),
@@ -52,6 +55,8 @@ class ManageTokensFragment :
             }
         }
 
+        manageTokensSwitchZeroBalances.setOnClickListener { viewModel.zeroBalancesClicked() }
+
         manageTokensList.setHasFixedSize(true)
         manageTokensList.adapter = tokensAdapter
 
@@ -62,6 +67,8 @@ class ManageTokensFragment :
 
         manageTokensSearch.requestFocus()
         manageTokensSearch.content.showSoftKeyboard()
+
+        manageTokensSwitchZeroBalances.bindFromMap(NonZeroBalanceFilter, viewModel.filtersEnabledMap, viewLifecycleOwner.lifecycleScope)
     }
 
     override fun inject() {
