+++ feature-assets/src/main/java/io/novafoundation/nova/feature_assets/presentation/tokens/manage/di/ManageTokensModule.kt
@@ -6,8 +6,11 @@ import androidx.lifecycle.ViewModelProvider
 import dagger.Module
 import dagger.Provides
 import dagger.multibindings.IntoMap
+import io.novafoundation.nova.common.di.scope.ScreenScope
 import io.novafoundation.nova.common.di.viewmodel.ViewModelKey
 import io.novafoundation.nova.common.di.viewmodel.ViewModelModule
+import io.novafoundation.nova.feature_assets.data.repository.assetFilters.AssetFiltersRepository
+import io.novafoundation.nova.feature_assets.domain.assets.filters.AssetFiltersInteractor
 import io.novafoundation.nova.feature_assets.domain.tokens.manage.ManageTokenInteractor
 import io.novafoundation.nova.feature_assets.presentation.AssetsRouter
 import io.novafoundation.nova.feature_assets.presentation.tokens.manage.ManageTokensViewModel
@@ -15,6 +18,12 @@ import io.novafoundation.nova.feature_assets.presentation.tokens.manage.model.Mu
 
 @Module(includes = [ViewModelModule::class])
 class ManageTokensModule {
+
+    @Provides
+    @ScreenScope
+    fun provideAssetFiltersInteractor(
+        assetFiltersRepository: AssetFiltersRepository
+    ) = AssetFiltersInteractor(assetFiltersRepository)
 
     @Provides
     internal fun provideViewModel(fragment: Fragment, factory: ViewModelProvider.Factory): ManageTokensViewModel {
@@ -27,12 +36,14 @@ class ManageTokensModule {
     fun provideViewModel(
         router: AssetsRouter,
         interactor: ManageTokenInteractor,
-        commonUiMapper: MultiChainTokenMapper
+        commonUiMapper: MultiChainTokenMapper,
+        assetFiltersInteractor: AssetFiltersInteractor
     ): ViewModel {
         return ManageTokensViewModel(
             router = router,
             interactor = interactor,
-            commonUiMapper = commonUiMapper
+            commonUiMapper = commonUiMapper,
+            assetFiltersInteractor = assetFiltersInteractor
         )
     }
 }
