+++ feature-assets/src/main/res/layout/item_asset_header.xml
@@ -91,20 +91,10 @@
         android:layout_height="wrap_content"
         android:layout_marginEnd="8dp"
         android:src="@drawable/ic_search"
-        app:layout_constraintBottom_toBottomOf="@+id/balanceListFilters"
-        app:layout_constraintEnd_toStartOf="@+id/balanceListFilters"
-        app:layout_constraintTop_toTopOf="@+id/balanceListFilters"
-        app:tint="@color/chip_icon" />
-
-    <io.novafoundation.nova.common.view.IconButton
-        android:id="@+id/balanceListFilters"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginEnd="8dp"
-        android:src="@drawable/ic_chip_filter"
         app:layout_constraintBottom_toBottomOf="@+id/balanceListManage"
         app:layout_constraintEnd_toStartOf="@+id/balanceListManage"
-        app:layout_constraintTop_toTopOf="@+id/balanceListManage" />
+        app:layout_constraintTop_toTopOf="@+id/balanceListManage"
+        app:tint="@color/chip_icon" />
 
     <io.novafoundation.nova.common.view.IconButton
         android:id="@+id/balanceListManage"
